sun is the same in a relative way
