--- client/blocks/login/login-header.scss
@@ -1,3 +1,0 @@
-.login-header-text__client-name {
-	text-transform: capitalize;
-}
