+++ client/login/wp-login/style.scss
@@ -937,3 +937,7 @@ $image-height: 47px;
 :root {
 	--login-form-column-max-width: 327px;
 }
+
+.wp-login__login-block-header-client-name {
+	text-transform: capitalize;
+}
